[DataLayer]More changes in order to fix issues when exporting [known issue: invalid number even when normal format and valid number is given :ORA-01722]

+++ src/main/java/lapr/project/datalayer/DAONetworkAnalysis.java
@@ -52,14 +52,14 @@ public class DAONetworkAnalysis extends DAOManager {
         cs.setString(5, netAnalysis.getEndNode().getId());
         cs.setString(6, netAnalysis.getType());
         cs.setString(7, netAnalysis.getName());
-        cs.setFloat(8, (float) netAnalysis.getTravellTime());
-        cs.setFloat(9, (float) netAnalysis.getEnergyConsumption());
-        cs.setFloat(10, (float) netAnalysis.getAverageVelocity());
-        cs.setFloat(11, (float) netAnalysis.getDistance());
-        cs.setFloat(12, (float) netAnalysis.getTollCost());
-        cs.setFloat(13, (float) netAnalysis.getLoad());
-        cs.setFloat(14, (float) netAnalysis.getFuelMass());
-        cs.setFloat(15, (float) netAnalysis.getFuelVolume());
+        cs.setDouble(8, netAnalysis.getTravellTime());
+        cs.setDouble(9, netAnalysis.getEnergyConsumption());
+        cs.setDouble(10, netAnalysis.getAverageVelocity());
+        cs.setDouble(11, netAnalysis.getDistance());
+        cs.setDouble(12, netAnalysis.getTollCost());
+        cs.setDouble(13, netAnalysis.getLoad());
+        cs.setDouble(14, netAnalysis.getFuelMass());
+        cs.setDouble(15, netAnalysis.getFuelVolume());
         if (netAnalysis.getType().equalsIgnoreCase("Shortest Travel Time")) {
             cs.setString(16, "N/A");
             cs.setString(17, "N/A");
